First commit of files for OOTP analysis:
Extravaganza is the file used for yearly blog posts.

MVPNames was used to find top players in each league history to name the MVP awards.

RunExpectancy calculates run values for each baserunner-out state and then a table of the run value of each plate appearance outcome.

GroupPlayers is still under construction - building a way to group hitters and pitchers and find the best matchups.